+++ write_octal.c
@@ -13,7 +13,7 @@ size_t write_octal(va_list args, flag_t *flag, size_t length_modifier)
 	size_t nwrite = 0;
 	uintmax_t num = (length_modifier == SHORT)
 					? va_arg(args, unsigned int)
-					: va_arg(args, unsigned long int);
+					: va_arg(args, unsigned int);
 
 
 	if (flag->hash && num > 0)
